clients/py/requirements.txt
cbor2
pynacl
web3
